restructure type files for client
