+++ src/utilities/rocketLaunchLiveClient/types.ts
@@ -1,6 +1,4 @@
-import axios from "axios";
-
-type LaunchesParams = {
+export type LaunchesParams = {
   id?: string | string[];
   cospar_id?: string;
   after_date?: string;
@@ -102,7 +100,7 @@ export type Launch = {
   modified: string;
 };
 
-type LaunchesResponse = {
+export type LaunchesResponse = {
   valid_auth: boolean;
   count: number;
   limit: number;
@@ -110,30 +108,3 @@ type LaunchesResponse = {
   last_page: number;
   result: Launch[];
 };
-
-export default class RocketLaunchLiveClient {
-  private baseUrl: string = "https://fdo.rocketlaunch.live";
-
-  private fetcher = axios.create({
-    headers: {
-      Authorization: `Bearer ${process.env.RLL_KEY}`,
-    },
-  });
-
-  constructor() {
-    this.fetcher
-      .get("https://fdo.rocketlaunch.live/json/launches")
-      .then((res) => console.log(res.data.result));
-  }
-
-  public fetchLaunches(options: LaunchesParams) {
-    const url = new URL(this.baseUrl);
-    url.pathname = "/json/launches";
-    const params = new URLSearchParams(options);
-    url.search = params.toString();
-
-    return this.fetcher
-      .get<LaunchesResponse>(url.toString())
-      .then((res) => res.data.result);
-  }
-}
